# docs/swagger.yaml
definitions:
...
  inout.BaseResponse:
    properties:
      error_code:
        type: integer
      error_description:
        type: string
    type: object
...
  model.AuthProvider:
    enum:
    - email
    - gmail
    - apple
    type: string
    x-enum-varnames:
    - AuthProviderEmail
    - AuthProviderGmail
    - AuthProviderApple
...
  model.UserStatus:
    enum:
    - active
    - suspended
    - inactive
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusSuspended
    - UserStatusInactive
...
  user.AcceptInviteOut:
    properties:
      data:
        $ref: '#/definitions/user.AcceptInviteResult'
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  user.AcceptInviteResult:
    properties:
      message:
        type: string
      organization_id:
        type: string
      organization_name:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
  user.ActivityItem:
    properties:
      description:
        type: string
      id:
        type: string
      timestamp:
        type: string
      type:
        type: string
    type: object
  user.DashboardData:
    properties:
      organization_count:
        type: integer
      personal_projects:
        type: integer
      recent_activity:
        items:
          $ref: '#/definitions/user.ActivityItem'
        type: array
      user:
        $ref: '#/definitions/user.User'
    type: object
  user.DashboardOut:
    properties:
      data:
        $ref: '#/definitions/user.DashboardData'
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  user.DenyInviteOut:
    properties:
      data:
        $ref: '#/definitions/user.DenyInviteResult'
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  user.DenyInviteResult:
    properties:
      message:
        type: string
      organization_id:
        type: string
      organization_name:
        type: string
      status:
        type: string
    type: object
  user.Notification:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      message:
        type: string
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  user.NotificationsOut:
    properties:
      data:
        items:
          $ref: '#/definitions/user.Notification'
        type: array
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  user.PendingInvite:
    properties:
      expires_at:
        type: string
      id:
        type: string
      invited_at:
        type: string
      organization_id:
        type: string
      organization_name:
        type: string
      role:
        type: string
      token:
        type: string
    type: object
  user.PendingInvitesOut:
    properties:
      data:
        items:
          $ref: '#/definitions/user.PendingInvite'
        type: array
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
  user.User:
    properties:
      auth_provider:
        $ref: '#/definitions/model.AuthProvider'
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_email_verified:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      status:
        $ref: '#/definitions/model.UserStatus'
      updated_at:
        type: string
      username:
        type: string
    type: object
  user.UserOut:
    properties:
      data:
        $ref: '#/definitions/user.User'
      error_code:
        type: integer
      error_description:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/users/account:
    delete:
      consumes:
      - application/json
      description: Delete current user's account permanently
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - User Management
  /api/v1/users/dashboard:
    get:
      consumes:
      - application/json
      description: Get user dashboard with overview of projects and activities
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.DashboardOut'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get user dashboard
      tags:
      - User Management
  /api/v1/users/invites:
    get:
      consumes:
      - application/json
      description: Get all pending organization invitations for the current user
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.PendingInvitesOut'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get pending invitations
      tags:
      - User Management
  /api/v1/users/invites/{token}/accept:
    post:
      consumes:
      - application/json
      description: Accept an organization invitation using the invitation token
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AcceptInviteOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Accept organization invite
      tags:
      - User Management
  /api/v1/users/invites/{token}/deny:
    post:
      consumes:
      - application/json
      description: Decline an organization invitation using the invitation token
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.DenyInviteOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Deny organization invite
      tags:
      - User Management
  /api/v1/users/notifications:
    get:
      consumes:
      - application/json
      description: Get list of user notifications
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.NotificationsOut'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get user notifications
      tags:
      - User Management
  /api/v1/users/notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - User Management
  /api/v1/users/profile:
    get:
      consumes:
      - application/json
      description: Get current user's profile information
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserOut'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Update current user's profile information
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: User profile data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User Management
swagger: "2.0"